requirements.txt
python-decouple==3.6
urllib3==1.25.8
